root@8.130.122.185/requirements.txt
django==4.2.16
django-cors-headers==4.8.0
requests==2.32.4
gunicorn==23.0.0
PyPDF2==3.0.1
python-dotenv>=1.1.1
azure-storage-blob>=12.25.1
pycryptodome>=3.20.0
pymysql>=1.1.0